+++ packages/polythene-css-classes/button.js
@@ -2,6 +2,7 @@
 export default {
   base:       "pe-button",
   component:  "pe-button pe-text-button",
+  row:        "pe-button-row",
 
   // elements
   content:    "pe-button__content",